% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Makara.R
\name{Makara}
\alias{Makara}
\title{Makara R6 Class}
\description{
An R6 class representing a connection to the Makara database.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connection}}{A database connection object, created using a package such as \code{ROracle::dbConnect()}.}

\item{\code{table}}{A table or view name that contains the data.}

\item{\code{sql}}{A SQL statement to query the database (DB).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Makara-new}{\code{Makara$new()}}
\item \href{#method-Makara-print}{\code{Makara$print()}}
\item \href{#method-Makara-finalize}{\code{Makara$finalize()}}
\item \href{#method-Makara-clone}{\code{Makara$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Makara-new"></a>}}
\if{latex}{\out{\hypertarget{method-Makara-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Makara R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Makara$new(connection, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{DB connection}

\item{\code{table}}{Table or view name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Makara-print"></a>}}
\if{latex}{\out{\hypertarget{method-Makara-print}{}}}
\subsection{Method \code{print()}}{
Prints the Makara R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Makara$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Makara-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Makara-finalize}{}}}
\subsection{Method \code{finalize()}}{
What to do when you remove the Makara R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Makara$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Makara-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Makara-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Makara$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
