% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Occurrence.R
\name{Occurrence}
\alias{Occurrence}
\title{Occurrence R6 Class}
\description{
An R6 class representing acoustic / vocal occurrence data from Passive Acoustic Monitoring.
}
\section{Super class}{
\code{\link[PAMvieweR:Makara]{PAMvieweR::Makara}} -> \code{Occurrence}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{species_id}}{A vector of species IDs to filter the data, see data table \code{PAGROUP.S_SPECIES}.}

\item{\code{agg_data}}{The aggregated dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Occurrence-get_data}{\code{Occurrence$get_data()}}
\item \href{#method-Occurrence-filter_species}{\code{Occurrence$filter_species()}}
\item \href{#method-Occurrence-aggregate}{\code{Occurrence$aggregate()}}
\item \href{#method-Occurrence-print}{\code{Occurrence$print()}}
\item \href{#method-Occurrence-clone}{\code{Occurrence$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="PAMvieweR" data-topic="Makara" data-id="finalize"><a href='../../PAMvieweR/html/Makara.html#method-Makara-finalize'><code>PAMvieweR::Makara$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PAMvieweR" data-topic="Makara" data-id="initialize"><a href='../../PAMvieweR/html/Makara.html#method-Makara-initialize'><code>PAMvieweR::Makara$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-get_data}{}}}
\subsection{Method \code{get_data()}}{
Reads the data into R from the DB.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-filter_species"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-filter_species}{}}}
\subsection{Method \code{filter_species()}}{
Filters the data by species.
This method modifies the SQL statement to query the DB.
If changed, you must re-execute \code{Occurrence$get_data()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$filter_species(species_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species_id}}{Vector of species IDs to filter by (see data table \code{PAGROUP.S_SPECIES})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
Aggregates the data by grouping variable(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$aggregate(.column, .function, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.column}}{Column name of the data to be aggregated}

\item{\code{.function}}{Character string of the function to aggregate the data in \code{.column}

Available functions:
\itemize{
\item{'n'}{ Count unique values}
\item{'min'}{ Minimum value}
\item{'max'}{ Maximum value}
\item{'mean'}{ Average value}
\item{'median'}{ Median value}
}}

\item{\code{...}}{Grouping variable(s)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-print"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-print}{}}}
\subsection{Method \code{print()}}{
Prints the Occurrence R6 Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
