% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Occurrence.R
\name{Occurrence}
\alias{Occurrence}
\title{Occurrence R6 Class}
\description{
An R6 class representing acoustic / vocal occurrence data from Passive Acoustic Monitoring (PAM).
}
\section{Super class}{
\code{\link[PAMvieweR:Makara]{PAMvieweR::Makara}} -> \code{Occurrence}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{species_id}}{A vector of species IDs, see data table \code{PAGROUP.S_SPECIES}.}

\item{\code{pam_data}}{The dataset (may be filtered, grouped, and / or summarized).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Occurrence-new}{\code{Occurrence$new()}}
\item \href{#method-Occurrence-get_data}{\code{Occurrence$get_data()}}
\item \href{#method-Occurrence-where_species}{\code{Occurrence$where_species()}}
\item \href{#method-Occurrence-filter}{\code{Occurrence$filter()}}
\item \href{#method-Occurrence-group}{\code{Occurrence$group()}}
\item \href{#method-Occurrence-aggregate}{\code{Occurrence$aggregate()}}
\item \href{#method-Occurrence-mutate}{\code{Occurrence$mutate()}}
\item \href{#method-Occurrence-order}{\code{Occurrence$order()}}
\item \href{#method-Occurrence-reset}{\code{Occurrence$reset()}}
\item \href{#method-Occurrence-print}{\code{Occurrence$print()}}
\item \href{#method-Occurrence-finalize}{\code{Occurrence$finalize()}}
\item \href{#method-Occurrence-clone}{\code{Occurrence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-new"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Makara R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$new(connection, table, data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{DB connection}

\item{\code{table}}{Table or view name}

\item{\code{data}}{\code{tibble} or \code{data.frame} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-get_data}{}}}
\subsection{Method \code{get_data()}}{
Queries the DB using an SQL statement and fetches the data into R.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-where_species"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-where_species}{}}}
\subsection{Method \code{where_species()}}{
Modifies the where clause of the SQL statement used by \code{get_data()} to include species IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$where_species(species_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species_id}}{Vector of species IDs (see data table \code{PAGROUP.S_SPECIES})}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method modifies the SQL statement used to query the DB.
If changed, you must re-execute \code{Occurrence$get_data()}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-filter"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-filter}{}}}
\subsection{Method \code{filter()}}{
Filters the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$filter(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See \link[dplyr]{filter}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-group"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-group}{}}}
\subsection{Method \code{group()}}{
Groups the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$group(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See \link[dplyr]{group_by}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
Summarizes the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$aggregate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See \link[dplyr]{summarize}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-mutate}{}}}
\subsection{Method \code{mutate()}}{
Modifies the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$mutate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See \link[dplyr]{mutate}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-order"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-order}{}}}
\subsection{Method \code{order()}}{
Orders the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$order(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{See \link[dplyr]{arrange}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-reset}{}}}
\subsection{Method \code{reset()}}{
Resets the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-print"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-print}{}}}
\subsection{Method \code{print()}}{
Prints the Occurrence R6 Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-finalize}{}}}
\subsection{Method \code{finalize()}}{
Removes the Occurrence R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Occurrence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Occurrence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occurrence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
